# openapi/oh.yaml
openapi: 3.0.1
info:
  title: Open Hospital API Documentation
  description: Open Hospital API Documentation
  contact:
    name: ApiInfo.DEFAULT_CONTACT
  license:
    name: GPL-3.0 license
    url: https://github.com/informatici/openhospital-api?tab=GPL-3.0-1-ov-file#readme
  version: 0.1.0
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: [ ]
paths:
...
  /bills/{id}:
    get:
      tags:
        - Bills
      operationId: getBill
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillDTO"
      security:
        - bearerAuth: [ ]
    put:
      tags:
        - Bills
      operationId: updateBill
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FullBillDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FullBillDTO"
      security:
        - bearerAuth: [ ]
    delete:
      tags:
        - Bills
      operationId: deleteBill
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: boolean
      security:
        - bearerAuth: [ ]
...
  /bills:
    get:
      tags:
        - Bills
      operationId: searchBills_1
      parameters:
        - name: datefrom
          in: query
          required: true
          schema:
            type: string
        - name: dateto
          in: query
          required: true
          schema:
            type: string
        - name: patient_code
          in: query
          required: false
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BillDTO"
      security:
        - bearerAuth: [ ]
    post:
      tags:
        - Bills
      operationId: newBill
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FullBillDTO"
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FullBillDTO"
      security:
        - bearerAuth: [ ]
  /bills/search/by/payments:
    post:
      tags:
        - Bills
      operationId: searchBillsByPayments
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/BillPaymentsDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BillDTO"
      security:
        - bearerAuth: [ ]
  /bills/search/by/item:
    post:
      tags:
        - Bills
      operationId: searchBills
      parameters:
        - name: datefrom
          in: query
          required: true
          schema:
            type: string
        - name: dateto
          in: query
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BillItemsDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BillDTO"
      security:
        - bearerAuth: [ ]
...
  /bills/pending:
    get:
      tags:
        - Bills
      operationId: getPendingBills
      parameters:
        - name: patient_code
          in: query
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BillDTO"
      security:
        - bearerAuth: [ ]
  /bills/pending/affiliate:
    get:
      tags:
        - Bills
      operationId: getPendingBillsAffiliate
      parameters:
        - name: patient_code
          in: query
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BillDTO"
      security:
        - bearerAuth: [ ]
  /bills/payments:
    get:
      tags:
        - Bills
      operationId: searchBillsPayments
      parameters:
        - name: datefrom
          in: query
          required: true
          schema:
            type: string
        - name: dateto
          in: query
          required: true
          schema:
            type: string
        - name: patient_code
          in: query
          required: false
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BillPaymentsDTO"
      security:
        - bearerAuth: [ ]
  /bills/payments/{bill_id}:
    get:
      tags:
        - Bills
      operationId: getPaymentsByBillId
      parameters:
        - name: bill_id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BillPaymentsDTO"
      security:
        - bearerAuth: [ ]
  /bills/items:
    get:
      tags:
        - Bills
      operationId: getDistinctItems
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BillItemsDTO"
      security:
        - bearerAuth: [ ]
  /bills/items/{bill_id}:
    get:
      tags:
        - Bills
      operationId: getItems
      parameters:
        - name: bill_id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BillItemsDTO"
      security:
        - bearerAuth: [ ]
...
components:
  schemas:
...
    PatientDTO:
      required:
        - age
        - bloodType
        - city
        - fatherName
        - firstName
        - motherName
        - secondName
        - sex
      type: object
      properties:
        code:
          type: integer
          format: int32
          readOnly: true
        firstName:
          maxLength: 50
          type: string
          description: First name of the patient
          example: Mario
        secondName:
          maxLength: 50
          type: string
          description: Last name of the patient
          example: Rossi
        name:
          type: string
          readOnly: true
        birthDate:
          type: string
          description: Birth date
          example: 1979-05-01
        age:
          type: integer
          description: Age
          format: int32
          example: 40
        agetype:
          type: string
          description: Age type
          example: "null"
        sex:
          type: string
          description: Sex
          example: M
          enum:
            - M
            - F
        address:
          maxLength: 50
          type: string
          description: Address
          example: "Via Roma, 12"
        city:
          maxLength: 50
          type: string
          description: City
          example: Verona
        nextKin:
          type: string
          readOnly: true
        telephone:
          maxLength: 50
          type: string
          description: Telephone
          example: "+393456789012"
        note:
          maxLength: 65535
          type: string
          description: Note
          example: Test insert new patient
        motherName:
          maxLength: 50
          type: string
          description: Mother's name
          example: Roberta
        mother:
          type: string
          description: "Mother's status (D=dead, A=alive)"
          example: A
          enum:
            - D
            - A
        fatherName:
          maxLength: 50
          type: string
          description: Father's name
          example: Giuseppe
        father:
          type: string
          description: "Father's status (D=dead, A=alive)"
          example: D
          enum:
            - D
            - A
        bloodType:
          type: string
          description: "Blood type (0-/+, A-/+ , B-/+, AB-/+)"
          example: A+
          enum:
            - 0-
            - 0+
            - A-
            - A+
            - B-
            - B+
            - AB-
            - AB+
        hasInsurance:
          type: string
          description: "HasInsurance (Y=Yes, N=no)"
          example: "N"
          enum:
            - "Y"
            - "N"
        parentTogether:
          type: string
          description: "Parent together (Y=Yes, N=no)"
          example: "N"
          enum:
            - "Y"
            - "N"
        taxCode:
          maxLength: 30
          type: string
          description: Tax code
          example: RSSMRA79E01L781N
        lock:
          type: integer
          description: Lock
          format: int32
          example: 0
        blobPhoto:
          type: array
          description: BlobPhoto
          items:
            type: string
            description: BlobPhoto
            format: byte
        hashCode:
          type: integer
          format: int32
          readOnly: true
        allergies:
          maxLength: 255
          type: string
          description: Allergies of patient
        anamnesis:
          maxLength: 255
          type: string
          description: Current anamnesis
        status:
          type: string
          description: Status
          example: I
          enum:
            - I
            - O
        consensusFlag:
          type: boolean
          description: Consensus flag
          example: true
        consensusServiceFlag:
          type: boolean
          description: Consensus service flag
          example: true
      description: Class representing a patient
...
    BillDTO:
      required:
        - amount
        - balance
        - date
        - listName
        - patName
        - patientTrue
        - status
        - update
        - user
      type: object
      properties:
        id:
          type: integer
          format: int32
        patient:
          $ref: "#/components/schemas/PatientDTO"
        listId:
          type: integer
          format: int32
        date:
          type: string
          description: Date of bill creation
          example: 2020-03-19T14:58:00.000Z
        update:
          type: string
          description: Date of bill updated
          example: 2020-03-19T14:58:00.000Z
        listName:
          maxLength: 50
          type: string
          description: Price list name
          example: Basic
        patientTrue:
          type: boolean
          description: Is bill belongs to a patient?
          example: true
        patName:
          maxLength: 100
          type: string
          description: Patient name
          example: Mario Rossi
        status:
          type: string
          description: Bill status
          example: O
        amount:
          type: number
          description: Bill Amount
          format: double
          example: 1000
        balance:
          type: number
          description: Bill balance
          format: double
          example: 1500
        user:
          type: string
          description: user name who create the bill
          example: admin
        lock:
          type: integer
          description: Lock
          format: int32
          example: 0
        list:
          type: boolean
      description: Class representing a bill
...
    BillItemsDTO:
      required:
        - itemAmount
        - itemDescription
        - itemDisplayCode
        - itemId
        - itemQuantity
        - priceId
      type: object
      properties:
        id:
          type: integer
          format: int32
        billId:
          type: integer
          description: Bill id
          format: int32
        priceId:
          type: string
          description: The price Id
          example: "104"
        itemDescription:
          type: string
          description: Item description
          example: Acetone 99 % 1ltr
        itemAmount:
          type: number
          description: Item amount
          format: double
          example: 1000
        itemQuantity:
          type: integer
          description: Item quantity
          format: int32
          example: 1
        itemDisplayCode:
          type: string
          description: Item display code
          example: Acetone
        itemId:
          type: string
          description: Item id
          example: "3"
        hashCode:
          type: integer
          format: int32
          readOnly: true
        price:
          type: boolean
      description: Class representing a billItem
    BillPaymentsDTO:
      required:
        - amount
        - billId
        - date
        - user
      type: object
      properties:
        id:
          type: integer
          format: int32
        billId:
          type: integer
          description: Bill id
          format: int32
        date:
          type: string
          description: Date of payment
          format: date-time
          example: 2020-03-19T14:58:00Z
        amount:
          type: number
          description: The payment amount
          format: double
          example: 500
        user:
          type: string
          description: The current user
          example: admin
        hashCode:
          type: integer
          format: int32
          readOnly: true
      description: Class representing a billPayment
    FullBillDTO:
      required:
        - bill
        - billItems
      type: object
      properties:
        bill:
          $ref: "#/components/schemas/BillDTO"
        billItems:
          type: array
          description: List of bill items elements
          items:
            $ref: "#/components/schemas/BillItemsDTO"
        billPayments:
          type: array
          description: List of bill payments elements
          items:
            $ref: "#/components/schemas/BillPaymentsDTO"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT